move document.jsx and fixes for linting
